+++ CHANGELOG.md
@@ -5,8 +5,10 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 ### Added
+- `debug` config flag
 
 ### Changed
+- Primary authentication is less hacky.
 
 ### Fixed